Create calculator.cpp
This program takes an operator and two operands from the user.

The operator is stored in variable op and two operands are stored in num1 and num2 respectively.

Then, switch...case statement is used for checking the operator entered by user.

If user enters + then, statements for case: '+' is executed and program is terminated.

If user enters - then, statements for case: '-' is executed and program is terminated.

This program works similarly for the * and / operators. But, if the operator doesn't matches any of the four character [ +, -, * and / ], the default statement is executed which displays error message.